requirements.txt
streamlit
pandas
numpy
plotly
groq
python-dotenv
reportlab
kaleido